pyproject.toml
[tool.poetry]
name = "ProjectZero"
version = "0.1.0"
description = "Django skeleton with most useful functionalities for new projects."
authors = ["Roman <98953084+Rokuflam@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
Django = "^5.0"
djangorestframework = "^3.14"
djangorestframework-simplejwt = "^5.3"
drf-spectacular = "^0.27"
gunicorn = "^21.2"
uvicorn = "^0.27"
pillow = "^10.2"
psycopg2-binary = "^2.9"
python = "^3.12"
sentry-sdk = {version = "^1.40", extras = ["django"]}
whitenoise = "^6.6"

[tool.poetry.dev-dependencies]
pylint = "^3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"